Restrict the TLS inline functions to inside the runtime, i.e. do not access the not-visible data from main.c The extra includes here are likely more than needed but not huge.
