+++ mono/mini/exceptions-mips.c
@@ -30,6 +30,7 @@
 #include "mini-mips.h"
 #include "mini-runtime.h"
 #include "aot-runtime.h"
+#include "mono/utils/mono-tls-inline.h"
 
 #define GENERIC_EXCEPTION_SIZE 256
