+++ mono/mini/interp/interp.c
@@ -25,6 +25,7 @@
 #include <mono/utils/gc_wrapper.h>
 #include <mono/utils/mono-math.h>
 #include <mono/utils/mono-counters.h>
+#include "mono/utils/mono-tls-inline.h"
 
 #ifdef HAVE_ALLOCA_H
 #   include <alloca.h>
