+++ mono/mini/mini-sparc.c
@@ -37,6 +37,7 @@
 #include "cpu-sparc.h"
 #include "jit-icalls.h"
 #include "ir-emit.h"
+#include "mono/utils/mono-tls-inline.h"
 
 /*
  * Sparc V9 means two things:
