+++ mono/mini/tramp-arm.c
@@ -31,6 +31,7 @@
 #ifndef DISABLE_INTERPRETER
 #include "interp/interp.h"
 #endif
+#include "mono/utils/mono-tls-inline.h"
 
 void
 mono_arch_patch_callsite (guint8 *method_start, guint8 *code_ptr, guint8 *addr)
