+++ mono/mini/tramp-arm64.c
@@ -25,7 +25,7 @@
 #ifndef DISABLE_INTERPRETER
 #include "interp/interp.h"
 #endif
-
+#include "mono/utils/mono-tls-inline.h"
 
 void
 mono_arch_patch_callsite (guint8 *method_start, guint8 *code_ptr, guint8 *addr)
