+++ mono/mini/tramp-ppc.c
@@ -24,6 +24,7 @@
 #include "mini.h"
 #include "mini-ppc.h"
 #include "mini-runtime.h"
+#include "mono/utils/mono-tls-inline.h"
 
 #if 0
 /* Same as mono_create_ftnptr, but doesn't require a domain */
