+++ mono/mini/tramp-sparc.c
@@ -20,6 +20,7 @@
 #include "mini.h"
 #include "mini-sparc.h"
 #include "jit-icalls.h"
+#include "mono/utils/mono-tls-inline.h"
 
 /*
  * mono_arch_get_unbox_trampoline:
